% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae_plot.R
\name{ae_plot}
\alias{ae_plot}
\title{Create AE Plot}
\usage{
ae_plot(data, footnote)
}
\arguments{
\item{data}{AE data frame}

\item{paramVar}{AE Preferred Term Column AETERM}

\item{aeStartVar}{AE Start day column}

\item{aeEndVar}{AE End day column}

\item{colorVar}{AE Severity}
}
\value{
an AE  plot created with ggplot
}
\description{
Create AE Plot
}
\examples{
 ae_plot(
  data = safetyCharts::stack_events() \%>\%
    mutate(seq = dplyr::row_number()) \%>\%
    filter(id == "01-717-1004"),
  footnote = "fff"
)

}
