% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae_plot_server.R
\name{ae_plot_server}
\alias{ae_plot_server}
\title{Event Timeline Module: Server}
\usage{
ae_plot_server(id, params, current_id)
}
\arguments{
\item{id}{\code{character} Module ID}

\item{params}{\code{list} Named list with \code{data} and \code{settings} (reactive)}

\item{current_id}{\code{character} Current participant ID (reactive)}
}
\value{
\code{function} Module server
}
\description{
Event Timeline Module: Server
}
