% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae_plot_ui.R
\name{ae_plot_UI}
\alias{ae_plot_UI}
\title{UI that facilitates the AE plot tab that includes AE plot and AE table outputs}
\usage{
ae_plot_UI(id)
}
\arguments{
\item{id}{module id}
}
\value{
AE plot tab UI
}
\description{
UI that facilitates the AE plot tab that includes AE plot and AE table outputs
}
