% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae_plot.R
\name{AEplot}
\alias{AEplot}
\title{Create AEplot}
\usage{
AEplot(dataCombined, eventVar, startDayVar, colorVar, endDayVar)
}
\arguments{
\item{colorVar}{AE Severity}

\item{data}{AE data frame}

\item{paramVar}{AE Preferred Term Column AETERM}

\item{aeStartVar}{AE Start day column}

\item{aeEndVar}{AE End day column}
}
\value{
an AE  plot created with ggplot
}
\description{
Create AEplot
}
\examples{
AEplot(
  data = safetyData::adam_adae,
  paramVar = "AETERM",
  aeStartVar = "ASTDY",
  aeEndVar = "AENDY",
  colorVar = "AENDY"
)

}
