% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_domains.R
\name{combine_domains}
\alias{combine_domains}
\title{Combine AE and CM data}
\usage{
combine_domains(aes_data, cm_data, settings)
}
\arguments{
\item{aes_data}{AE data}

\item{cm_data}{CM data}

\item{settings}{safetyGraphics settings}
}
\value{
combined dataset with stacked AE and CM data
}
\description{
Combine AE and CM data
}
\examples{
combine_domains(
     aes_data = params()$data$aes,
     cm_data = params()$data$cm,
     settings = params()$settings
     )
}
