% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lb_react.R
\name{lb_react}
\alias{lb_react}
\title{Create safety line plot}
\usage{
lb_react(data, paramVar, visVar, adyVar, baseVar, avalVar, lowVar, highVar)
}
\arguments{
\item{data}{long data frame such as LB or VS}

\item{paramVar}{term column}

\item{visVar}{analysis visit column}

\item{adyVar}{analysis day column}

\item{baseVar}{baseline value column}

\item{avalVar}{analysis value column}

\item{lowVar}{lower limit column}

\item{highVar}{upper limit column}
}
\value{
an lineplot created with ggplot
}
\description{
Create safety line plot
}
\examples{
lb_react(
    data=safetyData::adam_adlb,
    paramVar = "PARAM",
    visVar = "AVISIT",
    adyVar = "ADY",
    baseVar = "BASE",
    avalVar = "AVAL",
    lowVar = "A1LO",
    highVar = "A1HI"
   )


}
