% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lb_tbl.R
\name{lb_tbl}
\alias{lb_tbl}
\title{Create safety line plot}
\usage{
lb_tbl(data, paramVar, visVar, adyVar, avalVar, lowVar, highVar)
}
\arguments{
\item{data}{long data frame such as LB or VS}

\item{paramVar}{term column}

\item{visVar}{analysis visit column}

\item{adyVar}{analysis day column}

\item{avalVar}{analysis value column}

\item{lowVar}{lower limit column}

\item{highVar}{upper limit column}
}
\value{
an lineplot created with ggplot
}
\description{
Create safety line plot
}
\examples{
lb_tbl(
    data=safetyData::adam_adlbh,
    paramVar = "PARAM",
    visVar = "AVISIT",
    adyVar = "ADY",
    avalVar = "AVAL",
    lowVar = "A1LO",
    highVar = "A1HI"
   )


}
