% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lb_tbl_server.R
\name{lb_tbl_server}
\alias{lb_tbl_server}
\title{Lab Table Module: Server}
\usage{
lb_tbl_server(id, params, current_id)
}
\arguments{
\item{id}{\code{character} Shiny module ID}

\item{params}{\code{list} Named list with \code{data} and \code{settings} (reactive).}

\item{current_id}{\code{character} Current participant ID (reactive).}
}
\value{
Reactive containing AE reactable
}
\description{
Lab Table Module: Server
}
