% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lb_tbl_ui.R
\name{lb_tbl_ui}
\alias{lb_tbl_ui}
\title{Lab Table Module: UI}
\usage{
lb_tbl_ui(id)
}
\arguments{
\item{id}{\code{character} Module ID}
}
\value{
\code{list} Shiny tag list
}
\description{
Lab Table Module: UI
}
