% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview_server.R
\name{overview_server}
\alias{overview_server}
\title{Data Domain Listing Module: Server}
\usage{
overview_server(id, params, current_id)
}
\arguments{
\item{id}{\code{character} Shiny module ID}

\item{params}{\code{list} Named list with \code{data} and \code{settings} (reactive).}

\item{current_id}{\code{character} Current participant ID (reactive).}
}
\value{
\code{function} Module server
}
\description{
Data Domain Listing Module: Server
}
