% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview_ui.R
\name{overview_ui}
\alias{overview_ui}
\title{Data Domain Listing Module: UI}
\usage{
overview_ui(id)
}
\arguments{
\item{id}{\code{character} Module ID}
}
\value{
\code{list} Shiny tag list
}
\description{
Data Domain Listing Module: UI
}
