% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview_server.R
\name{OverviewServer}
\alias{OverviewServer}
\title{Server that facilitates the mapping of a single data element (column or field) with a simple select UI}
\usage{
OverviewServer(input, output, session, params, id)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}
}
\value{
A reactive containing the selected column
}
\description{
Server that facilitates the mapping of a single data element (column or field) with a simple select UI
}
