% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview_ui.R
\name{OverviewUI}
\alias{OverviewUI}
\title{UI that facilitates the mapping of a single data element (column or field) with a simple select UI}
\usage{
OverviewUI(id)
}
\arguments{
\item{id}{unique id for the UI}

\item{label}{label associated with the control}

\item{choices}{a list of options for the control}

\item{default}{default value for the control}
}
\value{
returns the selected value wrapped in a \code{reactive()}.
}
\description{
UI that facilitates the mapping of a single data element (column or field) with a simple select UI
}
