% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_server.R
\name{profile_server}
\alias{profile_server}
\title{Safety Profile Module: Server}
\usage{
profile_server(
  id,
  params,
  ptid = reactive({
     NULL
 })
)
}
\arguments{
\item{id}{\code{character} Shiny module ID}

\item{params}{\code{list} Named list with \code{data} and \code{settings} (reactive).}

\item{ptid}{\code{character} Current participant ID (reactive).}
}
\value{
\code{function} Module server
}
\description{
Safety Profile Module: Server
}
