% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profileApp.R
\name{profileApp}
\alias{profileApp}
\title{Safety Profile App}
\usage{
profileApp(
  data = list(dm = safetyData::adam_adsl, aes = safetyData::adam_adae, labs =
    safetyData::adam_adlbc, cm = safetyData::sdtm_cm),
  settings = NULL,
  ptid = NULL,
  runNow = TRUE
)
}
\arguments{
\item{data}{\code{list} Named list of data domains}

\item{settings}{\code{list} Named list of data domain metadata}

\item{ptid}{\code{character} Initial participant ID; defaults to first ID when sorted alphanumerically}

\item{runNow}{\code{logical} Run app immediately? Default: \code{TRUE}}
}
\description{
Safety Profile App
}
