% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/react_server.R
\name{react_server}
\alias{react_server}
\title{Safety Reactable Module - Server}
\usage{
react_server(id, params, current_id)
}
\arguments{
\item{id}{Shiny module id}

\item{params}{parameters object with \code{data} and \code{settings} options.}

\item{current_id}{current selected id}

\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}
}
\value{
Reactive containing AE reactable
}
\description{
Safety Reactable Module - Server
}
