% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safety_lineplot.R
\name{safety_lineplot}
\alias{safety_lineplot}
\title{Create safety line plot}
\usage{
safety_lineplot(data, paramVar, adyVar, avalVar)
}
\arguments{
\item{data}{long data frame such as LB or VS}

\item{paramVar}{term column}

\item{adyVar}{analysis day column}

\item{avalVar}{analysis value column}

\item{colorVar}{optional color variable}
}
\value{
an lineplot created with ggplot
}
\description{
Create safety line plot
}
\examples{
safety_lineplot(
  data = safetyData::adam_advs,
  paramVar = "PARAM",
  adyVar = "ADY",
  avalVar = "AVAL",
  colorVar = "ATPT"
)

}
