% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safety_lineplot_server.R
\name{safety_lineplot_server}
\alias{safety_lineplot_server}
\title{Safety Line plot Module - Server}
\usage{
safety_lineplot_server(id, params, current_id)
}
\arguments{
\item{id}{Shiny module id}

\item{params}{parameters object with \code{data} and \code{settings} options.}

\item{current_id}{current selected id}

\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}
}
\value{
Reactive containing AE plot and listing
}
\description{
Safety Line plot Module - Server
}
