% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safety_lineplot_ui.R
\name{safety_lineplot_UI}
\alias{safety_lineplot_UI}
\title{UI that facilitates the Safety line plot tab that includes safety line plot and corresponding table outputs}
\usage{
safety_lineplot_UI(id)
}
\arguments{
\item{id}{module id}
}
\value{
LB plot tab UI
}
\description{
UI that facilitates the Safety line plot tab that includes safety line plot and corresponding table outputs
}
