% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae_plot.R
\name{ae_plot}
\alias{ae_plot}
\title{Event Timeline}
\usage{
ae_plot(data, footnote = "")
}
\arguments{
\item{data}{\code{data.frame} Adverse events data domain}

\item{footnote}{\code{character} Footnote text}
}
\value{
\code{htmlwidget} Event timeline \code{plotly} object
}
\description{
Event Timeline
}
\examples{
data <- safetyCharts::stack_events() \%>\%
  dplyr::filter(
    !(is.na(.data$stdy) & is.na(.data$endy))
  ) \%>\%
  dplyr::arrange(
    .data$id, .data$stdy, .data$endy
  ) \%>\%
  dplyr::group_by(.data$id) \%>\%
  dplyr::mutate(
    seq = dplyr::row_number()
  ) \%>\%
  dplyr::ungroup()

data \%>\%
  dplyr::filter(
    .data$id == sample(unique(.data$id), 1)
  ) \%>\%
  ae_plot()

}
