% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae_plot_server.R
\name{ae_plot_server}
\alias{ae_plot_server}
\title{Title}
\usage{
ae_plot_server(input, output, session, params, id)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}

\item{params}{parameters object with \code{data} and \code{settings} options.}

\item{id}{Shiny output object}
}
\value{
Reactive containing AE plot and listing
}
\description{
Title
}
