% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae_plot_server.R
\name{ae_plot_server}
\alias{ae_plot_server}
\title{AE Plot Module - Server}
\usage{
ae_plot_server(id, params, current_id)
}
\arguments{
\item{id}{Shiny module id}

\item{params}{parameters object with \code{data} and \code{settings} options.}

\item{current_id}{current selected id}

\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}
}
\value{
Reactive containing AE plot and listing
}
\description{
AE Plot Module - Server
}
