% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae_plot_ui.R
\name{ae_plot_ui}
\alias{ae_plot_ui}
\title{Event Timeline Module: UI}
\usage{
ae_plot_ui(id)
}
\arguments{
\item{id}{\code{character} Module ID}
}
\value{
\code{list} Shiny tag list
}
\description{
Event Timeline Module: UI
}
