% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AEplot.R
\name{AEplot}
\alias{AEplot}
\title{Create AEplot}
\usage{
AEplot(data, paramVar, aeStartVar, aeEndVar, colorVar)
}
\arguments{
\item{data}{AE data frame}

\item{paramVar}{AE Preferred Term Column AETERM}

\item{aeStartVar}{AE Start day column}

\item{aeEndVar}{AE End day column}

\item{colorVar}{AE Severity}
}
\value{
an AE  plot created with ggplot
}
\description{
Create AEplot
}
\examples{
AEplot(
    data=safetyData::adam_adae,
    paramVar = "AETERM",
    aeStartVar= "ASTDY",
    aeEndVar="AENDY",
    colorVar="AENDY"
   )


}
