% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae_plot.R
\name{AEplot}
\alias{AEplot}
\title{Create AEplot}
\usage{
AEplot(data, footnote)
}
\arguments{
\item{data}{AE data frame}

\item{paramVar}{AE Preferred Term Column AETERM}

\item{aeStartVar}{AE Start day column}

\item{aeEndVar}{AE End day column}

\item{colorVar}{AE Severity}
}
\value{
an AE  plot created with ggplot
}
\description{
Create AEplot
}
\examples{
 AEplot(
  data = safetyCharts::stack_events() \%>\%
    mutate(seq = dplyr::row_number()) \%>\%
    filter(id == "01-717-1004"),
  footnote = "fff"
)
}
