% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_server.R
\name{id_server}
\alias{id_server}
\title{ID Select: Server}
\usage{
id_server(
  id,
  params,
  ptid = reactive({
     NULL
 })
)
}
\arguments{
\item{id}{Shiny module id}

\item{params}{parameters object with \code{data} and \code{settings} options. {reactive}}

\item{ptid}{ID to select when module is initialized {reactive}}
}
\value{
returns shiny module Server function
}
\description{
ID Select: Server
}
