% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lb_tbl.R
\name{lb_tbl}
\alias{lb_tbl}
\title{Lab Table}
\usage{
lb_tbl(data, measureVar, visitVar, studyDayVar, resultVar, llnVar, ulnVar)
}
\arguments{
\item{data}{long data frame such as LB or VS}

\item{measureVar}{term column}

\item{visitVar}{analysis visit column}

\item{studyDayVar}{analysis day column}

\item{resultVar}{analysis value column}

\item{llnVar}{lower limit column}

\item{ulnVar}{upper limit column}
}
\value{
an lineplot created with ggplot

\code{htmlwidget} Event timeline \code{plotly} object
}
\description{
Lab Table
}
\examples{
data <- safetyData::adam_adlbc \%>\%
  dplyr::filter(
    .data$USUBJID == sample(unique(.data$USUBJID), 1)
  )

data \%>\%
  lb_tbl(
    measureVar = "PARAM",
    visitVar = "AVISIT",
    studyDayVar = "ADY",
    resultVar = "AVAL",
    llnVar = "A1LO",
    ulnVar = "A1HI"
  )

}
