% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/react_ui.R
\name{lb_tbl_UI}
\alias{lb_tbl_UI}
\title{UI that facilitates the Safety reactable tab that includes safety reactable}
\usage{
lb_tbl_UI(id)
}
\arguments{
\item{id}{module id}
}
\value{
LB reactable tab UI
}
\description{
UI that facilitates the Safety reactable tab that includes safety reactable
}
