% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview_server.R
\name{overview_server}
\alias{overview_server}
\title{Server that facilitates the mapping of a single data element (column or field) with a simple select UI}
\usage{
overview_server(id, params, current_id)
}
\arguments{
\item{id}{Shiny module id}

\item{params}{parameters object with \code{data} and \code{settings} options.}

\item{current_id}{current selected id}

\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}
}
\value{
A reactive containing the selected column
}
\description{
Server that facilitates the mapping of a single data element (column or field) with a simple select UI
}
