% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_server.R
\name{profile_server}
\alias{profile_server}
\title{Safety Profile Module - Server}
\usage{
profile_server(id, params)
}
\arguments{
\item{id}{Shiny module id}

\item{params}{parameters object with \code{data} and \code{settings} options.}

\item{input}{module input}

\item{output}{module output}

\item{session}{module session}
}
\value{
returns shiny module Server function
}
\description{
Safety Profile Module - Server
}
