% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_ui.R
\name{profile_ui}
\alias{profile_ui}
\title{Safety Profile Module: UI}
\usage{
profile_ui(id)
}
\arguments{
\item{id}{\code{character} Module ID}
}
\value{
\code{list} Shiny tag list
}
\description{
Safety Profile Module: UI
}
