% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profileApp.R
\name{profileApp}
\alias{profileApp}
\title{Safety Profile App}
\usage{
profileApp(
  data = list(aes = safetyData::adam_adae, dm = safetyData::adam_adsl, labs =
    safetyData::adam_adlbc \%>\% filter(PARAMCD \%in\% c("PHOS", "GLUC", "PROT")), cm =
    safetyData::sdtm_cm),
  settings = NULL,
  ptid = NULL,
  runNow = TRUE
)
}
\arguments{
\item{settings}{safetyGraphics settings}

\item{ptid}{participant ID to select when app is initialized}

\item{dfAE}{AE Data}

\item{dfDemog}{demog data}
}
\description{
Safety Profile App
}
